+++ types/index.d.ts
@@ -1,4 +1,4 @@
-// Type definitions for Kyanite v1.0.2
+// Type definitions for Kyanite v1.5.0
 // Project: Kyanite
 // Definitions by: Dustin Hershman <dustinh17@gmail.com>
 
@@ -9,10 +9,7 @@ declare namespace K {
 
   type Ord = number | string | boolean | Date;
 
-  interface KeyValuePair<K, V> extends Array<K | V> {
-    0: K;
-    1: V;
-  }
+  type KeyValuePair<K, V> = [K, V];
 
   interface Schema {
     [key: string]: ((value: any) => any) | Schema;
@@ -109,9 +106,9 @@ declare namespace K {
     /**
      * Validates the same value when passed into two seperate functions
      */
-    both(pred1: Pred, pred2: Pred, data: any): boolean;
-    both(pred1: Pred, pred2: Pred): (data: any) => boolean;
-    both(pred1: Pred): (pred2: Pred) => (data: any) => boolean;
+    both<T = any>(pred1: Pred, pred2: Pred, data: T): boolean;
+    both(pred1: Pred, pred2: Pred): <T = any>(data: T) => boolean;
+    both(pred1: Pred): (pred2: Pred) => <T = any>(data: T) => boolean;
 
     /**
      * Takes 3 functions and a value, runs the functions in an if else setup, if the first function passes the second will run, otherwise the thrid will run
@@ -368,6 +365,12 @@ declare namespace K {
     fold<T = any, U = any, TResult = any>(fn: (a: T, acc: TResult) => TResult): (arr: ReadonlyArray<T>) => U;
 
     /**
+     * Takes an array of arrays which contain key value pairs and builds a new object
+     */
+    fromPairs<V>(pairs: Array<KeyValuePair<string, V>>): { [index: string]: V };
+    fromPairs<V>(pairs: Array<KeyValuePair<number, V>>): { [index: number]: V };
+
+    /**
      * Takes a needle and searches the haystack for the matching string
      */
     fuzzySearch(needle: string, haystack: string): boolean;
@@ -601,6 +604,12 @@ declare namespace K {
      */
     omit<T, E extends string>(keys: ReadonlyArray<E>, obj: T): Omit<T, E>;
     omit<E extends string>(keys: ReadonlyArray<E>): <T>(obj: T) => Omit<T, E>;
+
+    /**
+     * Builds out a new object but omits the key values from the new object that do NOT pass the predicate function
+     */
+    omitBy<T, E extends string>(fn: Pred, obj: T): Omit<T, E>;
+    omitBy<E extends string>(fn: Pred): <T>(obj: T) => Omit<T, E>;
 
     /**
      * Applies the second function to the values passed in, and then runs the first function against those new values,